Code clean
git-svn-id: http://pww.svn.pic.philips.com/DI-dev/trunk/Projects/PurAir/Android@3373 b40d9911-a802-4213-86a8-5840401a66a1

+++ PurAir/src/com/philips/cl/di/dev/pa/screens/MapOdActivity.java
@@ -112,7 +112,7 @@ public class MapOdActivity extends FragmentActivity {
 
 	    //"- ((paint.descent() + paint.ascent()) / 2)" is the distance from the baseline to the center.
 	    //int yPos = (int) ((canvas.getHeight() / 2) - ((paint.descent() + paint.ascent()) / 2)) ;  
-	    int yPos = (int) (canvas.getHeight()) - 5 ;  
+	    int yPos = (canvas.getHeight()) - 5 ;  
 	    canvas.drawText(text, xPos, yPos, paint);
 
 	    return  bm;
